requirements.txt
Flask
pymongo[srv]
python-dotenv
pydantic